+++ cfme/fixtures/pytest_selenium.py
@@ -24,7 +24,7 @@ from selenium.webdriver.common.by import By
 from selenium.webdriver.support import expected_conditions
 from selenium.webdriver.support.wait import WebDriverWait
 from selenium.webdriver.remote.webelement import WebElement
-from selenium.webdriver.support.select import Select
+from selenium.webdriver.support.select import Select as SeleniumSelect
 from multimethods import singledispatch, multidispatch
 
 import pytest
@@ -716,13 +716,38 @@ def set_text(loc, text):
         send_keys(el, text)
 
 
-def make_it_select(o):
-    """Make sure the object is select"""
-    if isinstance(o, Select):
-        move_to_element(o._el)
-        return o
-    else:
-        return Select(move_to_element(o))
+class Select(SeleniumSelect, object):
+    """ A proxy class for the real selenium Select() object.
+
+    We differ in one important point, that we can instantiate the object
+    without it being present on the page. The object is located at the beginning
+    of each function call.
+
+    Args:
+        loc: A locator.
+
+    Returns: A :py:class:`cfme.web_ui.Select` object.
+    """
+    def __init__(self, loc, multi=False):
+        if isinstance(loc, Select):
+            self._loc = loc._loc
+        else:
+            self._loc = loc
+        self.is_multiple = multi
+
+    @property
+    def _el(self):
+        return move_to_element(self)
+
+    def locate(self):
+        logger.error(type(self._loc))
+        return move_to_element(element(self._loc))
+
+    def observer_wait(self):
+        detect_observed_field(self._loc)
+
+    def __str__(self):
+        return "<%s.Select loc='%s'>" % (__name__, self._loc)
 
 
 @multidispatch
@@ -732,13 +757,13 @@ def select(loc, o):
 
 @select.method((object, ByValue))
 def _select_tuple(loc, val):
-    select_by_value(make_it_select(loc), val.value)
+    select_by_value(Select(loc), val.value)
 
 
 @select.method((object, basestring))
 @select.method((object, ByText))
 def _select_str(loc, s):
-    select_by_text(make_it_select(loc), str(s))
+    select_by_text(Select(loc), str(s))
 
 
 @select.method((object, Iterable))
@@ -806,13 +831,13 @@ def deselect(loc, o):
 
 @deselect.method((object, ByValue))
 def _deselect_val(loc, val):
-    deselect_by_value(make_it_select(loc), val.value)
+    deselect_by_value(Select(loc), val.value)
 
 
 @deselect.method((object, basestring))
 @deselect.method((object, ByText))
 def _deselect_text(loc, s):
-    deselect_by_text(make_it_select(loc), str(s))
+    deselect_by_text(Select(loc), str(s))
 
 
 @deselect.method((object, Iterable))
